Move and rename ReverseTransactions
Anything reusable in billing is held at the root of the folder. We change the name to try and make it clearer that this service is used in supplementary billing only.
